Mmarti (#126)
* parse assignment word

* split to files

* hotfix

* fix

* rm ft_realloc.c from interpretator

* fix

* another path to heredoc

* add var substitution, fix lp

* tilda substitution

* add pipes

* rm pipes

* rm pipes

* rm pipes
