+++ interpretator/Makefile
@@ -13,13 +13,17 @@
 srcs := parse.c\
 		err.c\
 		token_handle.c\
-		ft_realloc.c\
-		ft_strtok.c\
 		close_fds.c\
 		add_process.c\
 		fd_aggr.c\
-		redirect_list.c\
-		ft_free_jobs.c
+		word_list.c\
+		ft_free_jobs.c\
+		heredoc.c\
+		ft_open.c\
+		io_redir.c\
+		add_redir.c\
+		exec_jobs.c\
+		expansions.c
 
 objs := $(srcs:.c=.o)
