+++ interpretator/includes/interpretator.h
@@ -25,7 +25,7 @@
 # define DRRED_OPEN O_RDWR | O_CREAT | O_APPEND
 # define PERM_MODE S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH
 # define INT_LEN 10
-# define HEREDOC_FILE ".42sh_heredoc"
+# define HEREDOC_FILE "/tmp/.42sh_heredoc"
 # define INIT_AGGR_FD -5
 
 typedef int(*redirect_func)(t_lex_tkn **, t_process *, int);
@@ -36,6 +36,10 @@ typedef struct	s_open_files
 	size_t size;
 }				t_open_files;
 
+char		*substitution(char *str);
+void		exec_jobs(t_job *j);
+int			l_redir(t_lex_tkn **list, t_process *curr_proc, int io_number);
+int			here_doc(t_lex_tkn **list, t_process *curr, int io_number);
 int			l_aggr(t_lex_tkn **list, t_process *curr_proc, int io_number);
 int			ft_open(/*t_process *curr_proc,*/char *fname, int fl);
 int			g_aggr(t_lex_tkn **list, t_process *curr_proc, int io_number);
@@ -51,13 +55,12 @@ t_process	*add_process(void);
 int			syntax_err(t_lex_tkn *token);
 t_lex_tkn	**find_token(t_lex_tkn **list, int type);
 void		parse(t_lex_tkn **tokens);
-void		*ft_realloc(void *buf, size_t old, size_t new_size);
 char		**ft_strtok(char *s);
 char		*ft_strccut(char *str, char c);
 char		*ft_stricut(char *str, int i);
 t_lex_tkn	**split_list(t_lex_tkn **token);
 t_lex_tkn	**find_token(t_lex_tkn **list, int type);
 t_job		*job_new(void);
-int         redirect_list(t_lex_tkn **redir, t_process *cur_proc);
+int         word_list(t_lex_tkn **redir, t_process *cur_proc);
 
 #endif
