+++ interpretator/srcs/parse.c
@@ -19,7 +19,7 @@ int simp_command(t_lex_tkn **list)
 	if ((*list)->type == T_END)
 		return (0);
 	curr_proc = add_process();
-    return (redirect_list(list, curr_proc));
+    return (word_list(list, curr_proc));
 }
 
 int	pipe_sequence(t_lex_tkn **list)
@@ -47,7 +47,6 @@ int	logical_list(t_lex_tkn **list)
 	// tmp = find_token(list, LOGICAL);
 	// add_logical(tmp);
 	// tmp = split_list(tmp);
-	
 	if ((pipe_sequence(list)) < 0)
 		return (-1);
 	return (0);
@@ -103,30 +102,25 @@ void print_jobs(t_job *first_job)
 	}
 }
 
-int	get_token_ind(t_lex_tkn **token_list, t_lex_tkn *token)
-{
-	int i;
+// int	get_token_ind(t_lex_tkn **token_list, t_lex_tkn *token)
+// {
+// 	int i;
 
-	i = 0;
-	if (!token)
-		return (-1);
-	while (token_list[i] != token)
-		i++;
-	return(i);
-}
+// 	i = 0;
+// 	if (!token)
+// 		return (-1);
+// 	while (token_list[i] != token)
+// 		i++;
+// 	return(i);
+// }
 
-void	exec_jobs(t_job *j)
+void	ft_sub(t_lex_tkn **list)
 {
-	while (j)
+	while ((*list)->type != T_END)
 	{
-		if (!j->unactive)
-			launch_job(j, 0);
-		else 
-		{
-			if (j->err_message)
-				ft_putln(j->err_message);
-		}
-		j = j->next;
+		if ((*list)->type == T_WORD)
+			(*list)->value = substitution((*list)->value);
+		list++;
 	}
 }
 
@@ -134,6 +128,7 @@ void	parse(t_lex_tkn **tokens)
 {
 	if (!*tokens || (*tokens)->type == T_END)
 		return ;
+	ft_sub(tokens);
 	// lex_print_tkns(tokens);
 	// return;
 	if (start(tokens) == 0 && g_first_job)
