+++ jobs/srcs/launch_job.c
@@ -6,25 +6,28 @@ void launch_job(t_job *j, int foreground)
     pid_t       pid;
 
     p = j->first_process;
-    if (!p->args[0])
-        return ;
-    pid = make_process();
-    if (pid == 0)
-        launch_process(p, j->pgid, foreground);
-    else
-    {
-        p->pid = pid;
-        if (g_shell_is_interactive)
+    // while (p)
+    // {
+        pid = make_process();
+        if (pid == 0)
+            launch_process(p, j->pgid, foreground);
+        else
         {
-            if (!j->pgid)
+            p->pid = pid;
+            if (g_shell_is_interactive)
+            {
+                if (!j->pgid)
                 j->pgid = pid;
-            setpgid(pid, j->pgid);
+                setpgid(pid, j->pgid);
+            }
         }
-    }
-//    if (g_shell_is_interactive)
+        p = p->next;
+    // }
+    //waitpid(-1, 0, 0);
+    if (g_shell_is_interactive)
         wait_for_job(j);
-    // else if (foreground)
-    //     put_job_in_foreground(j, 0);
-    // else
-    //     put_job_in_background(j, 0);
+    else if (foreground)
+        put_job_in_foreground(j, 0);
+    else
+        put_job_in_background(j, 0);
 }
