+++ jobs/srcs/launch_process.c
@@ -44,7 +44,9 @@ void	launch_process(t_process *p, pid_t pgid, int foreground)
 		close(STDERR_FILENO);*/
 	}
 	redir(p->redir);
-	if (!(fname = (char *)get_hash_data(g_bin_table.table, p->args[0], g_bin_table.size)))
+	if (ft_strchr(p->args[0], '/'))
+		fname = p->args[0];
+	else if (!(fname = (char *)get_hash_data(g_bin_table.table, p->args[0], g_bin_table.size)))
 		err_exit("42sh", "command not found", p->args[0], NOERROR);
 	else if (execve(fname, p->args, g_env.env) < 0)
 		err_exit("42sh", "execve() error", p->args[0], NOERROR);
